Moved OrderTotal to presentational
