+++ client/src/components/CheckoutModal/index.js
@@ -5,7 +5,7 @@ import { Elements } from 'react-stripe-elements';
 
 import Modal from '../HOC/Modal';
 import StripeCheckoutForm from '../StripeCheckoutForm';
-import OrderTotal from '../OrderTotal';
+import OrderTotal from '../Presentational/OrderTotal';
 import TablesPageTitle from '../TablesPageTitle';
 import { Button } from '../../global-styles/styledComponents';
